Revert "combine the 2 analyzers into 1 for performance gains"
This reverts commit 02f107e99696846c91bc9ac4b01fbe9fcfb1790a.
