+++ src/main/java/org/javarosa/xpath/analysis/ContainsUncacheableExpressionAnalyzer.java
@@ -1,0 +1,34 @@
+package org.javarosa.xpath.analysis;
+
+import org.javarosa.xpath.expr.UncacheableXPathFuncExpr;
+import org.javarosa.xpath.expr.XPathFuncExpr;
+
+/**
+ * Created by amstone326 on 1/16/18.
+ */
+
+public class ContainsUncacheableExpressionAnalyzer extends XPathBooleanAnalyzer {
+
+    @Override
+    public void doAnalysis(XPathFuncExpr expr) {
+        if (expr instanceof UncacheableXPathFuncExpr) {
+            this.result = true;
+            this.shortCircuit = true;
+        }
+    }
+
+    @Override
+    protected boolean getDefaultValue() {
+        return false;
+    }
+
+    @Override
+    protected boolean aggregateResults() {
+        return orResults();
+    }
+
+    @Override
+    XPathAnalyzer initSameTypeAnalyzer() {
+        return new ContainsUncacheableExpressionAnalyzer();
+    }
+}
